Updated the default amount of parallel_calls to 1
With some comment on how to use it.
It will require some experimentation to see what is the fastest.

It is expected that for mu-auth; very small batch size and many
parallel calls work best.

If you go directly to virtuoso, which can max accept 10 calls; you
might want to limit the parallel_calls between 1 and 4.
But the batch sizes can be way bigger.
Note: for the latter; we'll have to beef up virtuoso.ini